refactor: Change mock nearOffers function location.
