+++ src/pages/offer-page/offer-page.tsx
@@ -1,10 +1,11 @@
 import { useParams } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
 import { OfferPreview } from '../../types/offer';
-import { getOfferPreviewById, getMockNearOfferPreviews } from './utils';
+import { getOfferPreviewById } from './utils';
 import { getRatingWidth } from '../../utils';
 import { getMockReviews } from '../../mock/reviews-mock';
 import { offerMock } from '../../mock/offer-mock';
+import { getMockNearOfferPreviews } from '../../mock/utils-mock';
 import Header from '../../components/header';
 import OfferGallery from '../../components/offer-gallery';
 import OfferFeatures from '../../components/offer-features';
